✨ [#182] Feat: msw 관심사 분리
