+++ src/shared/mocks/index.ts
@@ -1,0 +1,3 @@
+export * from './upload'
+export * from './product'
+export * from './review'
