+++ src/widgets/product/product-home/components/product-create-modal/modal/ProductCreateModal.stories.tsx
@@ -1,8 +1,7 @@
 import { Meta, StoryObj } from '@storybook/react'
-import { delay, http, HttpResponse, PathParams } from 'msw'
+import { uploadHandler, productCreateHandler } from '@shared/mocks'
 import { useToggle } from '@shared/hooks'
 import ProductCreateModal from './ProductCreateModal'
-import { ProductPayload } from '@shared/types'
 
 function ProductCreateModalWrapper() {
   const { isToggle, onOpenToggle, onCloseToggle } = useToggle()
@@ -11,32 +10,7 @@ function ProductCreateModalWrapper() {
 
 const meta: Meta<typeof ProductCreateModalWrapper> = {
   title: 'Components/Modal/ProductCreateModal',
-  parameters: {
-    msw: {
-      handlers: [
-        http.post(
-          `${process.env.NEXT_PUBLIC_BASE_URL}/images/upload`,
-          async () => {
-            await delay(1500)
-            return HttpResponse.json({ url: '성공적인 url 반환' })
-          },
-        ),
-        http.post<PathParams, ProductPayload>(
-          `${process.env.NEXT_PUBLIC_BASE_URL}/products`,
-          async ({ request }) => {
-            const data = await request.json()
-            await delay(1500)
-            return HttpResponse.json({
-              categoryId: 1,
-              image: data.image,
-              description: data.description,
-              name: data.name,
-            })
-          },
-        ),
-      ],
-    },
-  },
+  parameters: { msw: { handlers: [uploadHandler, productCreateHandler] } },
   component: ProductCreateModalWrapper,
 }
