+++ tsconfig.json
@@ -19,7 +19,6 @@
     "paths": {
       "@/*": ["./*"],
       "@app/*": ["./src/app/*"],
-      "@features/*": ["./src/features/*"],
       "@shared/*": ["./src/shared/*"],
       "@widgets/*": ["./src/widgets/*"]
     },
